+++ lib/Controller/LookupController.php
@@ -15,6 +15,7 @@ use OCA\FileChecksumSearch\Public\ChecksumApi;
 use OCA\FileChecksumSearch\Service\HashIndexService;
 use OCP\AppFramework\ApiController;
 use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\Attribute\ApiRoute;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
 use OCP\AppFramework\Http\DataResponse;
@@ -57,6 +58,7 @@ class LookupController
 	 */
 	#[NoAdminRequired]
 	#[NoCSRFRequired]
+	#[ApiRoute(verb: 'GET', url: '/api/1.0/lookup/{hash}')]
 	public function byHash(
 		string  $hash,
 		?string $algo = null,
@@ -107,6 +109,7 @@ class LookupController
 	 */
 	#[NoAdminRequired]
 	#[NoCSRFRequired]
+	#[ApiRoute(verb: 'GET', url: '/api/1.0/file/{fileId}/hashes')]
 	public function getHashesByFileId( int $fileId ): DataResponse
 	{
 
@@ -152,6 +155,7 @@ class LookupController
 	 */
 	#[NoAdminRequired]
 	#[NoCSRFRequired]
+	#[ApiRoute(verb: 'GET', url: '/api/1.0/file/{fileId}/duplicates')]
 	public function sameHash( int $fileId ): DataResponse
 	{
 
@@ -194,6 +198,7 @@ class LookupController
 	 * @noinspection PhpUnused
 	 */
 	#[NoAdminRequired]
+	#[ApiRoute(verb: 'POST', url: '/api/1.0/file/{fileId}/recalc')]
 	public function recalcHash(
 		int     $fileId,
 		?string $algo = null,