+++ src/duplicates.js
@@ -219,7 +219,7 @@ class DuplicateBrowser extends HTMLElement {
 				params.set( 'algo', this.#state.algo )
 			}
 
-			const url = OC.generateUrl( '/apps/file_checksum_search/api/v1/duplicates' ) + '?' + params.toString()
+			const url = OC.generateUrl( 'file_checksum_search.publicapi.findallduplicates' ) + '?' + params.toString()
 			const response = await fetch( url )
 			if ( !response.ok ) throw new Error( 'HTTP ' + response.status )
 			const data = await response.json()
@@ -332,7 +332,7 @@ class DuplicateBrowser extends HTMLElement {
 
 			for ( const file of group.files ) {
 				try {
-					const url = OC.generateUrl( '/apps/file_checksum_search/api/v1/file/{fileId}/recalc', {
+					const url = OC.generateUrl( 'file_checksum_search.publicapi.recalchash', {
 						fileId: file.fileid,
 					} ) + '?algo=' + group.algo
 					const res = await fetch( url, {